examples/custom_connector/none_auth_connector/requirements.txt
cozepy==0.13.0
Flask==2.2.5
requests-oauthlib==1.3.1
Werkzeug==2.2.3
python-dotenv==1.0.0
